Percentage modify Float to Double
git-svn-id: https://gsphere.genoscope.cns.fr/svn/NGL/trunk/ngl@3986 d0e871a8-0b98-4e59-9d77-5ea3c331444f

+++ app-ngl-sq/test/experiments/ExperimentTestHelper.java
@@ -59,7 +59,7 @@ public class ExperimentTestHelper {
 		exp.atomicTransfertMethods.put(1, atomicTransfert2);
 		
 		ContainerUsed container1_1=ExperimentTestHelper.getContainerUsed("CONTAINER1_1");
-		container1_1.percentage=(float) 20;
+		container1_1.percentage=20.0;
 		Content content1_1=new Content("CONTENT1_1","TYPE","CATEGORIE");
 		container1_1.contents=new ArrayList<Content>();
 		content1_1.properties=new HashMap<String, PropertyValue>();
@@ -70,7 +70,7 @@ public class ExperimentTestHelper {
 		container1_1.contents.add(content1_1);
 		
 		ContainerUsed container1_2=ExperimentTestHelper.getContainerUsed("CONTAINER1_2");
-		container1_2.percentage=(float) 80;
+		container1_2.percentage= 80.0;
 		Content content1_2=new Content("CONTENT1_2","TYPE","CATEGORIE");
 		container1_2.contents=new ArrayList<Content>();
 		content1_2.properties=new HashMap<String, PropertyValue>();
@@ -82,7 +82,7 @@ public class ExperimentTestHelper {
 		atomicTransfert1.inputContainerUseds.add(container1_2);
 		
 		ContainerUsed container2_2=ExperimentTestHelper.getContainerUsed("CONTAINER2_2");
-		container2_2.percentage=(float) 100;
+		container2_2.percentage= 100.0;
 		Content content2_2=new Content("CONTENT2_2","TYPE","CATEGORIE");
 		container2_2.contents=new ArrayList<Content>();
 		content2_2.properties=new HashMap<String, PropertyValue>();
